chore: add yellow (warning) to palette
