+++ sdk/src/commonMain/kotlin/ai/botstacks/sdk/ui/views/OnlineStatusIndicator.kt
@@ -5,14 +5,11 @@ import ai.botstacks.sdk.ui.BotStacks
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.dp
 
 @Composable
@@ -29,7 +26,7 @@ fun OnlineStatusIndicator(
                     OnlineStatus.Away -> BotStacks.colors.Dark._100
                     OnlineStatus.DND -> BotStacks.colorScheme.background
                     OnlineStatus.Offline -> BotStacks.colors.Error._500
-                    OnlineStatus.Online -> BotStacks.colors.Green._800
+                    OnlineStatus.Online -> BotStacks.colors.Success._800
                     OnlineStatus.UNKNOWN__ -> BotStacks.colorScheme.background
                 }, CircleShape
             ).then(modifier)
